+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ConcordTaskIT.java
@@ -251,4 +251,72 @@ public class ConcordTaskIT extends AbstractServerIT {
         byte[] ab = getLog(pir.getLogFileName());
         assertLog(".*Child process.*FAILED.*BOOOM.*", ab);
     }
+
+    @Test(timeout = 60000)
+    public void testForkWithItemsWithOutVariable() throws Exception {
+        String orgName = "org_" + randomString();
+
+        OrganizationsApi orgApi = new OrganizationsApi(getApiClient());
+        orgApi.createOrUpdate(new OrganizationEntry().setName(orgName));
+
+        String projectName = "project_" + randomString();
+
+        ProjectsApi projectsApi = new ProjectsApi(getApiClient());
+        projectsApi.createOrUpdate(orgName, new ProjectEntry()
+                .setName(projectName)
+                .setVisibility(ProjectEntry.VisibilityEnum.PUBLIC)
+                .setAcceptsRawPayload(true));
+
+        byte[] payload = archive(ProcessRbacIT.class.getResource("concordTaskForkWithItemsWithOut").toURI());
+        Map<String, Object> input = new HashMap<>();
+        input.put("archive", payload);
+        input.put("org", orgName);
+        input.put("project", projectName);
+
+        StartProcessResponse spr = start(input);
+
+        ProcessApi processApi = new ProcessApi(getApiClient());
+        ProcessEntry pir = waitForStatus(processApi, spr.getInstanceId(), ProcessEntry.StatusEnum.FAILED);
+
+        // ---
+
+        byte[] ab = getLog(pir.getLogFileName());
+        assertLog(".*color=RED.*", ab);
+        assertLog(".*color=WHITE.*", ab);
+        assertLog(".*Done.*\\[\\[.*\\], \\[.*\\]\\] is completed.*", ab);
+    }
+
+    @Test(timeout = 60000)
+    public void testForkWithItems() throws Exception {
+        String orgName = "org_" + randomString();
+
+        OrganizationsApi orgApi = new OrganizationsApi(getApiClient());
+        orgApi.createOrUpdate(new OrganizationEntry().setName(orgName));
+
+        String projectName = "project_" + randomString();
+
+        ProjectsApi projectsApi = new ProjectsApi(getApiClient());
+        projectsApi.createOrUpdate(orgName, new ProjectEntry()
+                .setName(projectName)
+                .setVisibility(ProjectEntry.VisibilityEnum.PUBLIC)
+                .setAcceptsRawPayload(true));
+
+        byte[] payload = archive(ProcessRbacIT.class.getResource("concordTaskForkWithItems").toURI());
+        Map<String, Object> input = new HashMap<>();
+        input.put("archive", payload);
+        input.put("org", orgName);
+        input.put("project", projectName);
+
+        StartProcessResponse spr = start(input);
+
+        ProcessApi processApi = new ProcessApi(getApiClient());
+        ProcessEntry pir = waitForStatus(processApi, spr.getInstanceId(), ProcessEntry.StatusEnum.FAILED);
+
+        // ---
+
+        byte[] ab = getLog(pir.getLogFileName());
+        assertLog(".*color=RED.*", ab);
+        assertLog(".*color=WHITE.*", ab);
+        assertLog(".*Done.*\\[\\[.*\\], \\[.*\\]\\] is completed.*", ab);
+    }
 }